refactor(rag): split sync source wizard into per-step components
sync-source-wizard.tsx was 761 lines in one file - the whole
multi-step flow, its form state, the connector picker, the
integration reuse path and the schedule, in one place. A reviewer
could not read one step without the other four in view, and every
future step that reads a list would have to thread its own failure
flag through the same component (the argument #61 made about the
page this file mirrors).

Following the one-component-per-step shape PR #221 established next
door in components/rag/, each step moves to its own file:

- sync-source-clone-step.tsx      CloneStep
- sync-source-connector-step.tsx  ConnectorStep
- sync-source-configure-step.tsx  ConfigureStep
- sync-source-schedule-step.tsx   ScheduleStep (with SYNC_MODES and
  SCHEDULE_PRESETS, which only it reads)

The wizard keeps what genuinely spans the steps and hands it down as
props: mode/step flow control, the shared form, the header chrome and
step indicator, and the connectorsFailed / orgIntegrationsFailed
flags - still required booleans reaching the step that renders the
list, so a step handed an already-empty array cannot reintroduce #32.
It drops from 761 to 320 lines.

No behaviour change - the step markup and the wizard's render logic
are moved verbatim. Verified: the existing sync-source-wizard and
derived-state integration tests pass unchanged; make lint-frontend
clean; the i18n guard clean; make test-frontend-cov green (240 files,
3756 tests, statements/functions/lines 100%, branches 97.64%).
components/rag/** is outside the coverage include, so no new per-step
tests were added - matching #221, which extracted its rag components
under the same integration tests.
